[FIX] account_invoice_move_currency; journal entries to draft
Before this change the user was not able to set a journal entry to
draft. This was because we introduce this error solving a secondary
currency problem in invoices in this commit https://github.com/ingadhoc/account-invoicing/commit/087bfad77538d5b7db4802a3ca9cf82c582fa401 that was actually a fix for another problem.

closes ingadhoc/account-invoicing#125

X-original-commit: 1238d6b65548466b7e7bb4a774c1dab4814be751
Signed-off-by: Juan José Scarafía <jjs@adhoc.com.ar>
Signed-off-by: Katherine Zaoral <kz@adhoc.com.ar>